requirements.txt
Flask
Flask-Login
Flask-Bcrypt  # For hashing passwords
pymongo[srv]  # MongoDB driver
pyotp         # For 2FA OTP generation
qrcode[pil]   # For generating QR codes for 2FA setup
python-dotenv
boto3
gunicorn